Remove ImageTextureLoader
- Replace KTXError exception with a function
- Link Stb only from editor
