+++ editor/lib/quoll/editor/asset/Stb.cpp
@@ -1,4 +1,5 @@
 #include "quoll/core/Base.h"
 
+#define STB_IMAGE_IMPLEMENTATION
 #define STB_IMAGE_WRITE_IMPLEMENTATION
-#include <stb_image_write.h>
+#include "Stb.h"
