+++ editor/lib/quoll/editor/asset/Stb.h
@@ -1,0 +1,4 @@
+#pragma once
+
+#include <stb_image.h>
+#include <stb_image_write.h>
