+++ editor/lib/quoll/editor/asset/gltf/TinyGLTF.h
@@ -5,7 +5,7 @@
 #define TINYGLTF_NO_INCLUDE_JSON
 #define TINYGLTF_USE_CPP14
 
+#include "../Stb.h"
 #include <nlohmann/json.hpp>
-#include <stb_image.h>
-#include <stb_image_write.h>
 #include <tiny_gltf.h>
+
