+++ editor/lib/quoll/editor/ui/IconRegistry.h
@@ -1,8 +1,12 @@
 #pragma once
 
-#include "quoll/renderer/RenderStorage.h"
-#include "quoll/rhi/RenderDevice.h"
 #include "quoll/rhi/RenderHandle.h"
+
+namespace quoll {
+
+class RenderStorage;
+
+}
 
 namespace quoll::editor {
